# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "timesheet API",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "accessCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer"
          },
          "code": {
            "type": "string",
            "enum": [
              "VALIDATION_FAILED",
              "INVALID_CREDENTIALS",
              "ACCESS_TOKEN_INVALID",
              "REFRESH_TOKEN_INVALID",
              "REFRESH_TOKEN_REUSED",
              "REFRESH_TOKEN_ROTATED",
              "CSRF_ORIGIN_REJECTED",
              "CSRF_TOKEN_INVALID",
              "EMAIL_TAKEN",
              "TOO_MANY_REQUESTS",
              "FORBIDDEN",
              "NOT_FOUND",
              "NO_STORE",
              "ALREADY_IN_STORE",
              "INVITE_INVALID",
              "ALREADY_CLOCKED_IN",
              "NOT_CLOCKED_IN",
              "REQUEST_PENDING",
              "REQUEST_CLOSED",
              "LEAVE_OVERLAP",
              "SUBSTITUTE_NOT_ACCEPTED",
              "INTERNAL"
            ]
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "statusCode",
          "code",
          "message"
        ]
      },
...
      "MyMembershipDto": {
        "type": "object",
        "properties": {
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "hourlyWage": {
            "type": "integer",
            "minimum": 0,
            "maximum": 1000000
          },
          "weeklyHours": {
            "type": "number",
            "minimum": 0,
            "maximum": 52
          },
          "workDaysPerWeek": {
            "type": "integer",
            "minimum": 1,
            "maximum": 7
          },
          "store": {
            "$ref": "#/components/schemas/StoreDto"
          }
        },
        "required": [
          "role",
          "hourlyWage",
          "weeklyHours",
          "workDaysPerWeek",
          "store"
        ]
      },
      "Role": {
        "type": "string",
        "enum": [
          "master",
          "member"
        ]
      },
      "StoreDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "lat": {
            "type": [
              "number",
              "null"
            ]
          },
          "lng": {
            "type": [
              "number",
              "null"
            ]
          },
          "fivePlus": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "lat",
          "lng",
          "fivePlus"
        ]
      },
      "CreateStoreBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateStoreBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50
          },
          "lat": {
            "type": [
              "number",
              "null"
            ],
            "minimum": -90,
            "maximum": 90
          },
          "lng": {
            "type": [
              "number",
              "null"
            ],
            "minimum": -180,
            "maximum": 180
          },
          "fivePlus": {
            "type": "boolean"
          }
        }
      },
      "InviteDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "used",
              "expired",
              "revoked"
            ]
          },
          "usedByUserId": {
            "type": [
              "string",
              "null"
            ]
          },
          "usedByNickname": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "id",
          "code",
          "createdAt",
          "expiresAt",
          "status",
          "usedByUserId",
          "usedByNickname"
        ]
      },
      "RedeemBody": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 32
          }
        },
        "required": [
          "code"
        ]
      },
      "MemberDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "nickname": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "hourlyWage": {
            "type": "integer",
            "minimum": 0,
            "maximum": 1000000
          },
          "weeklyHours": {
            "type": "number",
            "minimum": 0,
            "maximum": 52
          },
          "workDaysPerWeek": {
            "type": "integer",
            "minimum": 1,
            "maximum": 7
          },
          "joinedAt": {
            "type": "string"
          }
        },
        "required": [
          "userId",
          "nickname",
          "email",
          "role",
          "hourlyWage",
          "weeklyHours",
          "workDaysPerWeek",
          "joinedAt"
        ]
      },
      "UpdateMemberBody": {
        "type": "object",
        "properties": {
          "hourlyWage": {
            "type": "integer",
            "minimum": 0,
            "maximum": 1000000
          },
          "weeklyHours": {
            "type": "number",
            "minimum": 0,
            "maximum": 52
          },
          "workDaysPerWeek": {
            "type": "integer",
            "minimum": 1,
            "maximum": 7
          }
        }
      },
      "ShiftDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "start": {
            "type": "string"
          },
          "end": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "id",
          "userId",
          "start",
          "end"
        ]
      },
      "UpdateShiftBody": {
        "type": "object",
        "properties": {
          "start": {
            "type": "string",
            "format": "date-time"
          },
          "end": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        },
        "required": [
          "start",
          "end"
        ]
      },
      "DashboardDto": {
        "type": "object",
        "properties": {
          "store": {
            "$ref": "#/components/schemas/StoreDto"
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MemberDto"
            }
          },
          "shifts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ShiftDto"
            }
          },
          "absences": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AbsenceDto"
            }
          },
          "pendingRequests": {
            "type": "integer"
          }
        },
        "required": [
          "store",
          "members",
          "shifts",
          "absences",
          "pendingRequests"
        ]
      },
      "AbsenceDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "date": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "enum": [
              "paid_leave",
              "unpaid_leave",
              "substitution"
            ]
          },
          "sourceId": {
            "type": "string"
          }
        },
        "required": [
          "userId",
          "date",
          "kind",
          "sourceId"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/api/stores": {
      "post": {
        "summary": "매장 만들기 (마스터가 된다)",
        "security": [
          {
            "accessCookie": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "x-csrf-token",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStoreBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyMembershipDto"
                }
              }
            }
          },
          "409": {
            "description": "ALREADY_IN_STORE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/stores/me": {
      "get": {
        "summary": "내 매장",
        "security": [
          {
            "accessCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreDto"
                }
              }
            }
          },
          "404": {
            "description": "NO_STORE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "매장 설정 (마스터)",
        "security": [
          {
            "accessCookie": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "x-csrf-token",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStoreBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/stores/me/invites": {
      "get": {
        "summary": "초대 코드 목록 (마스터)",
        "security": [
          {
            "accessCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InviteDto"
                  }
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "초대 코드 발급 (마스터)",
        "security": [
          {
            "accessCookie": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "x-csrf-token",
            "in": "header"
          }
        ],
        "responses": {
          "201": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/stores/me/invites/{inviteId}": {
      "delete": {
        "summary": "초대 코드 취소 (마스터)",
        "security": [
          {
            "accessCookie": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "inviteId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "x-csrf-token",
            "in": "header"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/invites/redeem": {
      "post": {
        "summary": "초대 코드 등록 (멤버가 된다)",
        "security": [
          {
            "accessCookie": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "x-csrf-token",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RedeemBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyMembershipDto"
                }
              }
            }
          },
          "400": {
            "description": "INVITE_INVALID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "ALREADY_IN_STORE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "TOO_MANY_REQUESTS",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/stores/me/members": {
      "get": {
        "summary": "멤버 목록 (마스터)",
        "security": [
          {
            "accessCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MemberDto"
                  }
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/stores/me/members/{userId}": {
      "patch": {
        "summary": "멤버 급여 조건 (마스터)",
        "security": [
          {
            "accessCookie": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "userId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "x-csrf-token",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemberBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "멤버 내보내기 (마스터)",
        "security": [
          {
            "accessCookie": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "userId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "x-csrf-token",
            "in": "header"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/stores/me/members/{userId}/shifts": {
      "get": {
        "summary": "멤버 근무 기록 (마스터)",
        "security": [
          {
            "accessCookie": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "userId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "required": true,
            "name": "from",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "required": true,
            "name": "to",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ShiftDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/stores/me/shifts/{shiftId}": {
      "patch": {
        "summary": "근무 기록 수정 (마스터)",
        "security": [
          {
            "accessCookie": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "shiftId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "x-csrf-token",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateShiftBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShiftDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "근무 기록 삭제 (마스터)",
        "security": [
          {
            "accessCookie": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "shiftId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "x-csrf-token",
            "in": "header"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/stores/me/dashboard": {
      "get": {
        "summary": "대시보드 (마스터)",
        "security": [
          {
            "accessCookie": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-(0[1-9]|1[0-2])$"
            },
            "required": true,
            "name": "month",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}